Set up shared types package
