+++ backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/package.json
@@ -214,6 +214,7 @@
 		"spark-md5": "^3.0.2",
 		"ts-morph": "^25.0.1",
 		"undici": "5.22.1",
-		"uuid": "^11.1.0"
-	}
+                "uuid": "^11.1.0",
+                "@rocket-meals/shared": "file:../../../../shared"
+        }
 }
