+++ shared/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@rocket-meals/shared",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "index.js",
+  "types": "types.d.ts"
+}
